Added The Impostor Problem
We calculate the probability of somebody being the impostor a certain
number of times in a row during a game session of Among Us.

Visualisations were done with streamlit.

Renamed files so they were accessible to streamlit.
