+++ random_problems/python_teaching/Hangman.py
