+++ random_problems/special_times/special_time_generators.py
