+++ requirements.txt
@@ -1,7 +1,10 @@
 beautifulsoup4~=4.11.1
+graphviz~=0.20.1
 matplotlib~=3.5.1
 numpy~=1.22.3
 pandas~=1.4.2
+pygraphviz~=1.9
 requests~=2.27.1
 scikit-learn~=1.0.2
 seaborn~=0.11.2
+streamlit~=1.14.1
